aded import matplotlib and added test_findroots code into test_calculus.py
