+++ test_calculus.py
@@ -432,3 +432,76 @@ def test_calculate_square():
     # Test an invalid input (negative number)
     result = calculus.calculate_square(-4.0)
     assert math.isnan(result), "calculate_square(-4.0) should return NAN for invalid input."
+
+# Test for func_1_safe
+def test_func_1_safe():
+    """Test the func_1_safe function."""
+    assert abs(calc.func_1_safe(1)) - 1.557 < 1e-2
+    with pytest.raises(ValueError):
+        calc.func_1_safe(math.pi)
+
+# Test for func_2 (tanh)
+def test_func_2():
+    """Test the func_2 function (tanh)."""
+    assert abs(calc.func_2(0)) - 0.0 < 1e-6
+    assert abs(calc.func_2(1)) - 0.7616 < 1e-4
+    assert abs(calc.func_2(-1)) - 0.7616 < 1e-5
+
+# Test for func_3 (sin)
+def test_func_3():
+    """Test the func_3 function (sin)."""
+    assert abs(calc.func_3(0)) - 0.0 < 1e-6
+    assert abs(calc.func_3(math.pi / 2)) - 1.0 < 1e-6
+    assert abs(calc.func_3(math.pi)) - 0.0 < 1e-6
+
+# Test for calculate_accuracy
+def test_calculate_accuracy():
+    """Test the calculate_accuracy function."""
+    assert calc.calculate_accuracy(1.0, 1.0) == 0  # Exact match returns 0 accuracy
+    assert calc.calculate_accuracy(1.00001, 1.0) == 4
+
+# Test for apply_methods
+def test_apply_methods():
+    """Test the apply_methods function."""
+    try:
+        calc.apply_methods(calc.func_2, (-1, 1), "y(x) = tanh(x)", true_root=0.0, filename="test_plot.png")
+    except ValueError as e:
+        pytest.fail(f"apply_methods failed with ValueError: {str(e)}")
+    except FileNotFoundError as e:
+        pytest.fail(f"apply_methods failed due to missing file: {str(e)}")
+    except TypeError as e:
+        pytest.fail(f"apply_methods failed due to TypeError: {str(e)}")
+    except AssertionError as e:  # Specific exception for failed assertions
+        pytest.fail(f"apply_methods failed due to AssertionError: {str(e)}")
+    except Exception as e:  # Catching any unexpected errors
+        pytest.fail(f"apply_methods failed with an unexpected error: {str(e)}")
+
+# Test for plot_function_with_roots
+def test_plot_function_with_roots():
+    """Test the plot_function_with_roots function."""
+    try:
+        calc.plot_function_with_roots(calc.func_3, (0, math.pi), [math.pi], 'test_plot.png', "y(x) = sin(x)")
+    except ValueError as e:
+        pytest.fail(f"plot_function_with_roots failed with ValueError: {str(e)}")
+    except FileNotFoundError as e:
+        pytest.fail(f"plot_function_with_roots failed due to missing file: {str(e)}")
+    except TypeError as e:
+        pytest.fail(f"plot_function_with_roots failed due to TypeError: {str(e)}")
+    except AssertionError as e:  # Specific exception for failed assertions
+        pytest.fail(f"plot_function_with_roots failed due to AssertionError: {str(e)}")
+    except Exception as e:  # Catching any unexpected errors
+        pytest.fail(f"plot_function_with_roots failed with an unexpected error: {str(e)}")
+
+# Test for find_roots
+def test_find_roots():
+    """Test the find_roots function."""
+    try:
+        calc.find_roots()
+    except ValueError as e:
+        pytest.fail(f"find_roots failed with ValueError: {str(e)}")
+    except TypeError as e:
+        pytest.fail(f"find_roots failed with TypeError: {str(e)}")
+    except AssertionError as e:  # Specific exception for failed assertions
+        pytest.fail(f"find_roots failed due to AssertionError: {str(e)}")
+    except Exception as e:  # Catching any unexpected errors
+        pytest.fail(f"find_roots failed with an unexpected error: {str(e)}")
